Move admin activity aggregators to data package
