--- admin/tabs/activity/aggregators/index.js
@@ -1,3 +1,0 @@
-export * from './ChildCountAggregator';
-export * from './LeafCountAggregator';
-export * from './RangeAggregator';
