+++ data/cube/aggregate/ChildCountAggregator.js
@@ -5,18 +5,12 @@
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 
-import {Aggregator} from '@xh/hoist/data/cube/aggregate/Aggregator';
+import {Aggregator} from './Aggregator';
 
-export class LeafCountAggregator extends Aggregator {
+export class ChildCountAggregator extends Aggregator {
 
     aggregate(rows, fieldName) {
-        let count = 0;
-
-        for (const row of rows) {
-            count += row.isLeaf ? 1 : row.data[fieldName];
-        }
-
-        return count;
+        return rows.length;
     }
 
     replace(rows, currAgg, update) {
