+++ data/index.js
@@ -21,9 +21,12 @@ export * from './filter/Utils';
 export * from './cube/aggregate/Aggregator';
 export * from './cube/aggregate/AverageAggregator';
 export * from './cube/aggregate/AverageStrictAggregator';
+export * from './cube/aggregate/ChildCountAggregator';
+export * from './cube/aggregate/LeafCountAggregator';
 export * from './cube/aggregate/MaxAggregator';
 export * from './cube/aggregate/MinAggregator';
 export * from './cube/aggregate/NullAggregator';
+export * from './cube/aggregate/RangeAggregator';
 export * from './cube/aggregate/SingleAggregator';
 export * from './cube/aggregate/SumAggregator';
 export * from './cube/aggregate/SumStrictAggregator';
